Implement auth

+++ backend/controllers/goalController.mjs
@@ -5,7 +5,7 @@ import Goal from "../models/goalModel.mjs";
 // @route GET /api/goals
 // @access Private
 const getGoals = asyncHandler(async (req, res) => {
-	const goals = await Goal.find();
+	const goals = await Goal.find({ user: req.user });
 	res.json(goals);
 });
 
@@ -17,7 +17,7 @@ const setGoal = asyncHandler(async (req, res) => {
 		res.status(400);
 		throw new Error("Please add a text field");
 	}
-	const goal = await Goal.create({ text: req.body.text });
+	const goal = await Goal.create({ text: req.body.text, user: req.user });
 	res.json(goal);
 });
 
@@ -29,6 +29,10 @@ const updateGoal = asyncHandler(async (req, res) => {
 	if (!goal) {
 		res.status(400);
 		throw new Error("Goal not found");
+	}
+	if (goal.user.toString() !== req.user.id) {
+		res.status(401);
+		throw new Error("Unauthorised");
 	}
 	const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, { new: true });
 	res.json(updatedGoal);
@@ -42,6 +46,10 @@ const deleteGoal = asyncHandler(async (req, res) => {
 	if (!goal) {
 		res.status(400);
 		throw new Error("Goal not found");
+	}
+	if (goal.user.toString() !== req.user.id) {
+		res.status(401);
+		throw new Error("Unauthorised");
 	}
 	await Goal.findByIdAndDelete(req.params.id);
 	res.status(200).send("Deleted");
